src/requirements.txt
# Core workflow orchestration
langgraph>=1.0.0

# LLM providers
openai>=1.0.0
anthropic>=0.20.0

# Configuration
python-dotenv>=1.0.0

# Data & scientific computing
numpy>=1.24
pandas>=2.0
scipy>=1.12

# Causal inference libraries
dowhy>=0.11
econml>=0.15

# Graph & visualization
networkx>=3.0
matplotlib>=3.8
graphviz>=0.20

# Jupyter
ipykernel>=6.0
jupyter>=1.0
